[REFACT] Isolate UI component into independent Composable Function
  - Isolate UI component into independent Composable Function (MainServiceScreen)
